First big fix done
1. Explain the motivation of the work more clearly. This is specifically the fact that it is a compilation that helps participants understand what work has been done and published.
a. Note that we need to figure out the TITLE, whether to say “IETF” as now or "IETF/IRTF/IAB” (as it is used in https://www.ietf.org/privacy-statement/), depending on the path of the references we have. We have references from IETF, from IRTF, and from IAB.
b. We need to make sure that we do not direct future work. That is, compile work but not say that this is used for future IETF work